Fixed warnings during compilation. (#25)
* Fixed warnings during TS3 plugin compilation (in 32 bit mode at least).

* Massive cleanup. Removed all warnings to acre.dll compilation.

* Also remove no longer active settings status in TS3 settings dialog.

+++ extensions/src/ACRE2Arma/common/arguments.hpp
@@ -41,7 +41,7 @@ namespace acre {
     public:
         arguments(const std::string & str) : _original(str), _internal_index(0) {
             _args = acre::split(str, ',');
-            for (int i = 0; i < _args.size(); i++) {
+            for (size_t i = 0; i < _args.size(); i++) {
                 _args[i] = trim(_args[i]);
             }
         }
